Continued cleaning up and added a few new features

+++ webservice/src/main/java/kickbackapp/Service/EventService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import kickbackapp.NotFoundException;
 import kickbackapp.Entity.EventEntity;
 import kickbackapp.Repository.EventRepo;
 
@@ -24,24 +25,35 @@ public class EventService {
         return eventRepo.findAll();
     }
     
-    public EventEntity findEvent(Integer id) {
+    public EventEntity findEvent(Integer id) throws NotFoundException {
     	EventEntity event = eventRepo.findById(id);
-    	return event;
+    	if (event != null) {
+    		return event;
+    	} else {
+    		throw new NotFoundException(String.format("No such event with id %d", id));
+    	}
     }
     
-
-    public EventEntity saveEvent(EventEntity event) {
+    /**
+     * Saves event with owner as owner
+     */
+    public EventEntity saveEvent(EventEntity event, String owner) {
     	System.out.println("saveEvent");
+    	event.setOwner(owner);
     	return eventRepo.save(event);
     }
     
-    
-    public void deleteEvent(Integer id) {
+    /**
+     * Deletes event if user is owner of event
+     */
+ 	public void deleteEvent(Integer id, String owner) {
     	System.out.println("find event for delete:"+id);
 		EventEntity event = eventRepo.findById(id);
 		if (event != null) {
-			System.out.println("deleting event:"+event.getName());
-			eventRepo.delete(event);
+			if (event.getOwner() ==  owner) {
+				System.out.println("deleting event:"+event.getName());
+				eventRepo.delete(event);
+			}
 		}	
     }
 }
